Dynamic chromosome and Fitness
