+++ src/GeneticSharp.Domain/Chromosomes/GenericChromosome.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using GeneticSharp.Domain.Fitnesses;
@@ -24,41 +23,5 @@ namespace GeneticSharp.Domain.Chromosomes
         {
             return GetGenes().Select(g => (TGene) g.Value);
         }
-    }
-
-    public class DynamicChromosome<TGene> : GenericChromosomeBase<TGene>
-    {
-
-
-        private Func<int, TGene> GeneGenerator { get; set; }
-
-
-        public DynamicChromosome(int length) : base(length)
-        {
-        }
-
-        public override IChromosome CreateNew()
-        {
-            return new DynamicChromosome<TGene>(this.Length){GeneGenerator = GeneGenerator};
-        }
-
-
-        protected override TGene GenerateNewGeneValue(int geneIndex)
-        {
-            return GeneGenerator(geneIndex);
-        }
-    }
-
-    public abstract class GenericFitness<TGene> : IFitness
-    {
-        public double Evaluate(IChromosome chromosome)
-        {
-            var typedChromosome = chromosome as GenericChromosomeBase<TGene>;
-            var typedGenes = typedChromosome.GetTypedGenes();
-            return EvaluateGenes(typedGenes);
-        }
-
-        protected abstract double EvaluateGenes(IEnumerable<TGene> genes);
-
     }
 }
